requirements.txt
fastapi
uvicorn
sqlalchemy
fastapi-users
python-dotenv
cloudinary
pytest
